Ships 5.0.1.1
Finally released

+++ src/MoseShipsBukkit/Utils/VesselLoader.java
@@ -47,7 +47,7 @@ public class VesselLoader {
 	@SuppressWarnings("deprecation")
 	public static boolean classicLoader(File file){
 		YamlConfiguration config = YamlConfiguration.loadConfiguration(file);
-		OfflinePlayer owner = Bukkit.getOfflinePlayer(config.getString("ShipsData.Player.Name"));
+		String sUUID = config.getString("ShipsData.Player.Name");
 		int max = config.getInt("ShipsData.Config.Block.Max");
 		int min = config.getInt("ShipsData.Config.Block.Min");
 		int engine = config.getInt("ShipsData.Config.Speed.Engine");
@@ -76,25 +76,28 @@ public class VesselLoader {
 					}
 				}
 				if (loc != null){
-					if (loc.getBlock().getState() instanceof Sign){
-						Sign sign = (Sign)loc.getBlock().getState();
-						if (sign.getLine(0).equals(ChatColor.YELLOW + "[Ships]")){
-							String typeString = sign.getLine(1).replace(ChatColor.BLUE + "", "");
-							String typeStringFix = typeString.replace("2", "");
-							VesselType type = VesselType.getTypeByName(typeStringFix);
-							if (type != null){
-								if (type instanceof ClassicVessel){
-								type.setDefaultSpeed(engine);
-								type.setMaxBlocks(max);
-								type.setMinBlocks(min);
-								Vessel vessel = new Vessel(sign, name, type, owner, loc);
-								((ClassicVessel)vessel.getVesselType()).loadVesselFromClassicFile(vessel, file);
-								file.delete();
-								vessel.save();
-								return true;
+					if (sUUID != null){
+						OfflinePlayer owner = Bukkit.getOfflinePlayer(sUUID);
+						if (loc.getBlock().getState() instanceof Sign){
+							Sign sign = (Sign)loc.getBlock().getState();
+							if (sign.getLine(0).equals(ChatColor.YELLOW + "[Ships]")){
+								String typeString = sign.getLine(1).replace(ChatColor.BLUE + "", "");
+								String typeStringFix = typeString.replace("2", "");
+								VesselType type = VesselType.getTypeByName(typeStringFix);
+								if (type != null){
+									if (type instanceof ClassicVessel){
+									type.setDefaultSpeed(engine);
+									type.setMaxBlocks(max);
+									type.setMinBlocks(min);
+									Vessel vessel = new Vessel(sign, name, type, owner, loc);
+									((ClassicVessel)vessel.getVesselType()).loadVesselFromClassicFile(vessel, file);
+									file.delete();
+									vessel.save();
+									return true;
+									}
+								}else{
+									Bukkit.getConsoleSender().sendMessage(ChatColor.RED + sign.getLine(1) + " is no longer supported in Ships 5");
 								}
-							}else{
-								Bukkit.getConsoleSender().sendMessage(ChatColor.RED + sign.getLine(1) + " is no longer supported in Ships 5");
 							}
 						}
 					}
@@ -108,8 +111,7 @@ public class VesselLoader {
 		YamlConfiguration config = YamlConfiguration.loadConfiguration(file);
 		ConsoleCommandSender sender = Bukkit.getConsoleSender();
 		try{
-			UUID uuid = UUID.fromString(config.getString("ShipsData.Player.Name"));
-			OfflinePlayer owner =  Bukkit.getOfflinePlayer(uuid);
+			String sUUID = config.getString("ShipsData.Player.Name");
 			String vesselTypeS = config.getString("ShipsData.Type");
 			int max = config.getInt("ShipsData.Config.Block.Max");
 			int min = config.getInt("ShipsData.Config.Block.Min");
@@ -119,47 +121,57 @@ public class VesselLoader {
 			List<String> uuidSL = config.getStringList("PlayerLocation");
 			String name = file.getName().replace(".yml", "");
 			if (vesselTypeS != null){
-				if (locS != null){
-					if (teleportS != null){
-						String[] locM = locS.split(",");
-						Location loc = new Location(Bukkit.getWorld(locM[3]), Double.parseDouble(locM[0]), Double.parseDouble(locM[1]), Double.parseDouble(locM[2]));
-						if (loc.getBlock().getState() instanceof Sign){
-							Sign sign = (Sign)loc.getBlock().getState();
-							String[] teleportM = teleportS.split(",");
-							Location teleport = new Location(Bukkit.getWorld(teleportM[3]), Double.parseDouble(teleportM[0]), Double.parseDouble(teleportM[1]), Double.parseDouble(teleportM[2]));
-							VesselType vesselType = VesselType.getTypeByName(vesselTypeS).clone();
-							if (vesselType != null){
-								Vessel vessel = new Vessel(sign, name, vesselType, owner, teleport, false);
-								vesselType.setDefaultSpeed(engine);
-								vesselType.setMaxBlocks(max);
-								vesselType.setMinBlocks(min);
-								if (uuidSL != null){
-									Map<OfflinePlayer, BlockVector> locationB = new HashMap<OfflinePlayer, BlockVector>();
-									for(String uuidS : uuidSL){
-										String[] uuidArgs = uuidS.split(",");
-										OfflinePlayer player2 = Bukkit.getOfflinePlayer(UUID.fromString(uuidArgs[0]));
-										int X = Integer.parseInt(uuidArgs[1]);
-										int Y = Integer.parseInt(uuidArgs[2]);
-										int Z = Integer.parseInt(uuidArgs[3]);
-										Block block = vessel.getSign().getBlock().getRelative(X, Y, Z);
-										BlockVector vector = new BlockVector(vessel.getSign().getBlock(), block);
-										locationB.put(player2, vector);
+				if(sUUID != null){
+					UUID uuid = UUID.fromString(sUUID);
+					OfflinePlayer owner =  Bukkit.getOfflinePlayer(uuid);
+					if (locS != null){
+						if (teleportS != null){
+							String[] locM = locS.split(",");
+							Location loc = new Location(Bukkit.getWorld(locM[3]), Double.parseDouble(locM[0]), Double.parseDouble(locM[1]), Double.parseDouble(locM[2]));
+							if (loc.getWorld() != null){
+								if (loc.getBlock().getState() instanceof Sign){
+									Sign sign = (Sign)loc.getBlock().getState();
+									String[] teleportM = teleportS.split(",");
+									Location teleport = new Location(Bukkit.getWorld(teleportM[3]), Double.parseDouble(teleportM[0]), Double.parseDouble(teleportM[1]), Double.parseDouble(teleportM[2]));
+									VesselType vesselType = VesselType.getTypeByName(vesselTypeS).clone();
+									if (vesselType != null){
+										Vessel vessel = new Vessel(sign, name, vesselType, owner, teleport, false);
+										vesselType.setDefaultSpeed(engine);
+										vesselType.setMaxBlocks(max);
+										vesselType.setMinBlocks(min);
+										if (uuidSL != null){
+											Map<OfflinePlayer, BlockVector> locationB = new HashMap<OfflinePlayer, BlockVector>();
+											for(String uuidS : uuidSL){
+												String[] uuidArgs = uuidS.split(",");
+												OfflinePlayer player2 = Bukkit.getOfflinePlayer(UUID.fromString(uuidArgs[0]));
+												int X = Integer.parseInt(uuidArgs[1]);
+												int Y = Integer.parseInt(uuidArgs[2]);
+												int Z = Integer.parseInt(uuidArgs[3]);
+												Block block = vessel.getSign().getBlock().getRelative(X, Y, Z);
+												BlockVector vector = new BlockVector(vessel.getSign().getBlock(), block);
+												locationB.put(player2, vector);
+											}
+										}
+										vesselType.loadVesselFromFiveFile(vessel, file);
+										vessel.save();
+										return true;
+									}else{
+										sender.sendMessage(Ships.runShipsMessage("Vessel Loader: " + name + " has not loaded. This is because " + vesselTypeS + " has not loaded into Ships VesselType data system. This is a programming issue", true));
 									}
+								}else{
+									sender.sendMessage(Ships.runShipsMessage("Vessel Loader: " + name + " sign can not be located", true));
 								}
-								vesselType.loadVesselFromFiveFile(vessel, file);
-								vessel.save();
-								return true;
 							}else{
-								sender.sendMessage(Ships.runShipsMessage("VesselType Loader: " + name + " has not loaded. This is because " + vesselTypeS + " has not loaded into Ships VesselType data system. This is a programming issue", true));
+								sender.sendMessage(Ships.runShipsMessage("Vessel Loader: " + name + " world not loaded. Contact MoseMister with your world loading plugin", true));
 							}
 						}else{
-							sender.sendMessage(Ships.runShipsMessage("VesselType Loader: " + name + " sign can not be located", true));
+							sender.sendMessage(Ships.runShipsMessage("Vessel Loader: " + name + " teleport location has failed to load. Check config for 'Teleport' co-ods", true));
 						}
 					}else{
-						sender.sendMessage(Ships.runShipsMessage("VesselType Loader: " + name + " teleport location has failed to load. Check config for 'Teleport' co-ods", true));
+						sender.sendMessage(Ships.runShipsMessage("Vessel Loader: " + name + " sign location has failed to load. Check config for 'Sign' co-ods", true));
 					}
 				}else{
-					sender.sendMessage(Ships.runShipsMessage("VesselType Loader: " + name + " sign location has failed to load. Check config for 'Sign' co-ods", true));
+					sender.sendMessage(Ships.runShipsMessage("Vessel Loader:" + name + " failed reading player", true));
 				}
 			}else{
 				sender.sendMessage(Ships.runShipsMessage("VesselType Loader: " + name + " VesselType has not been found in data file. Check config for 'Type'", true));
